Restructure root layout and utils
