+++ app/lib/utils.ts
@@ -74,154 +74,6 @@ export function isDiscounted(price: MoneyV2, compareAtPrice: MoneyV2) {
   return false;
 }
 
-function resolveToFromType(
-  {
-    customPrefixes,
-    pathname,
-    type,
-  }: {
-    customPrefixes: Record<string, string>;
-    pathname?: string;
-    type?: string;
-  } = {
-    customPrefixes: {},
-  },
-) {
-  if (!pathname || !type) return "";
-
-  /*
-    MenuItemType enum
-    @see: https://shopify.dev/api/storefront/unstable/enums/MenuItemType
-  */
-  const defaultPrefixes = {
-    BLOG: "blogs",
-    COLLECTION: "collections",
-    COLLECTIONS: "collections", // Collections All (not documented)
-    FRONTPAGE: "frontpage",
-    HTTP: "",
-    PAGE: "pages",
-    CATALOG: "collections/all", // Products All
-    PRODUCT: "products",
-    SEARCH: "search",
-    SHOP_POLICY: "policies",
-  };
-
-  const pathParts = pathname.split("/");
-  const handle = pathParts.pop() || "";
-  const routePrefix: Record<string, string> = {
-    ...defaultPrefixes,
-    ...customPrefixes,
-  };
-
-  switch (true) {
-    // special cases
-    case type === "FRONTPAGE":
-      return "/";
-
-    case type === "ARTICLE": {
-      const blogHandle = pathParts.pop();
-      return routePrefix.BLOG
-        ? `/${routePrefix.BLOG}/${blogHandle}/${handle}/`
-        : `/${blogHandle}/${handle}/`;
-    }
-
-    case type === "COLLECTIONS":
-      return `/${routePrefix.COLLECTIONS}`;
-
-    case type === "SEARCH":
-      return `/${routePrefix.SEARCH}`;
-
-    case type === "CATALOG":
-      return `/${routePrefix.CATALOG}`;
-
-    // common cases: BLOG, PAGE, COLLECTION, PRODUCT, SHOP_POLICY, HTTP
-    default:
-      return routePrefix[type]
-        ? `/${routePrefix[type]}/${handle}`
-        : `/${handle}`;
-  }
-}
-
-/*
-  Parse each menu link and adding, isExternal, to and target
-*/
-function parseItem(primaryDomain: string, env: Env, customPrefixes = {}) {
-  return (
-    item:
-      | MenuFragment["items"][number]
-      | MenuFragment["items"][number]["items"][number],
-  ):
-    | EnhancedMenu["items"][0]
-    | EnhancedMenu["items"][number]["items"][0]
-    | null => {
-    if (!item?.url || !item?.type) {
-      // eslint-disable-next-line no-console
-      console.warn("Invalid menu item.  Must include a url and type.");
-      return null;
-    }
-
-    // extract path from url because we don't need the origin on internal to attributes
-    const { host, pathname } = new URL(item.url);
-
-    const isInternalLink =
-      host === new URL(primaryDomain).host || host === env.PUBLIC_STORE_DOMAIN;
-
-    const parsedItem = isInternalLink
-      ? // internal links
-        {
-          ...item,
-          isExternal: false,
-          target: "_self",
-          to: resolveToFromType({ type: item.type, customPrefixes, pathname }),
-        }
-      : // external links
-        {
-          ...item,
-          isExternal: true,
-          target: "_blank",
-          to: item.url,
-        };
-
-    if ("items" in item) {
-      return {
-        ...parsedItem,
-        items: item.items
-          // @ts-ignore
-          .map(parseItem(primaryDomain, env, customPrefixes))
-          .filter(Boolean),
-      } as EnhancedMenu["items"][number];
-    }
-    return parsedItem as EnhancedMenu["items"][number]["items"][number];
-  };
-}
-
-/*
-  Recursively adds `to` and `target` attributes to links based on their url
-  and resource type.
-  It optionally overwrites url paths based on item.type
-*/
-export function parseMenu(
-  menu: MenuFragment,
-  primaryDomain: string,
-  env: Env,
-  customPrefixes = {},
-): EnhancedMenu | null {
-  if (!menu?.items) {
-    // eslint-disable-next-line no-console
-    console.warn("Invalid menu passed to parseMenu");
-    return null;
-  }
-
-  const parser = parseItem(primaryDomain, env, customPrefixes);
-
-  const parsedMenu = {
-    ...menu,
-    items: menu.items.map(parser).filter(Boolean),
-  } as EnhancedMenu;
-
-  return parsedMenu;
-}
-
 export const INPUT_STYLE_CLASSES =
   "appearance-none rounded dark:bg-transparent border focus:border-line focus:ring-0 w-full py-2 px-3 text-body placeholder:text-body leading-tight focus:shadow-outline";
